% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clema.R
\name{clema}
\alias{clema}
\title{Lemmatize words using a chilean spanish dictionary}
\usage{
clema(words)
}
\arguments{
\item{words}{A word or a vector of words to lemmatize}
}
\value{
A converted vector of words
}
\description{
Lemmatize words using a chilean spanish dictionary
}
\examples{
clema("carabinera")
}
