% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clema.R
\name{clema}
\alias{clema}
\title{Lemmatize words using a chilean spanish dictionary}
\usage{
clema(words, replace = TRUE)
}
\arguments{
\item{words}{A word or a vector of words to lemmatize}

\item{replace}{If true the function will replace the NA values with originals}
}
\value{
A lemmatized vector of words
}
\description{
Lemmatize words using a chilean spanish dictionary
}
\examples{
clema("carabinera")
}
